// src/app/core/models/strategy-management/kpi.model.ts
export interface KpiResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: Kpis[];
}

export interface Kpis {
   
}

export interface KpiReviewResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: Kpis[];
}

export interface Kpis {
   
}